Add caching

+++ src/main/java/com/moskalyuk/ispairbe/dto/News.java
@@ -4,18 +4,20 @@ import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.Date;
-import java.util.LinkedList;
-import java.util.Map;
+
 
 @Builder
 @Getter
 public class News {
-    Integer id;
+
+    Long id;
+
     String link;
+
     String title;
+
     LocalDateTime date;
+
     Integer counter;
-//    Map<String, LinkedList<String>> comments;
 
 }
